[DMS-1124] Slice 2: Deduplicate inlined-scope normalization into shared helper
CompiledScopeAdapterFactory and ScopeTopologyIndex each carried
an identical private NormalizeAdditionalScopes / CountScopeDepth
pair. Promote them to a shared InlinedScopeNormalization helper
in Backend.External.Profile so both callers (one in
Backend.External, one in Backend) resolve to the same
implementation. Behavior unchanged; 576/576 backend unit tests
pass.
