+++ src/dms/backend/EdFi.DataManagementService.Backend.External/Profile/CompiledScopeAdapterFactory.cs
@@ -37,7 +37,7 @@ public static class CompiledScopeAdapterFactory
             .TablePlansInDependencyOrder.Select(tp => tp.TableModel.JsonScope.Canonical)
             .ToHashSet(StringComparer.Ordinal);
 
-        var normalizedAdditionalScopes = NormalizeAdditionalScopes(additionalScopes, tableScopes);
+        var normalizedAdditionalScopes = InlinedScopeNormalization.Normalize(additionalScopes, tableScopes);
 
         // Build a lookup of JsonScope canonical string -> ScopeKind for parent resolution.
         // Starts with table-backed scopes; additional scopes are added below.
@@ -332,42 +332,4 @@ public static class CompiledScopeAdapterFactory
         // Collection columns: SourceJsonPath is already scope-relative with $ root marker
         return canonicalPath.StartsWith("$.", StringComparison.Ordinal) ? canonicalPath[2..] : canonicalPath;
     }
-
-    private static IReadOnlyList<(string JsonScope, ScopeKind Kind)> NormalizeAdditionalScopes(
-        IReadOnlyList<(string JsonScope, ScopeKind Kind)>? additionalScopes,
-        IReadOnlySet<string> knownScopes
-    )
-    {
-        if (additionalScopes is not { Count: > 0 })
-        {
-            return [];
-        }
-
-        List<(string JsonScope, ScopeKind Kind, int Depth, int Index)> normalized = [];
-        HashSet<string> seenScopes = new(knownScopes, StringComparer.Ordinal);
-
-        for (var index = 0; index < additionalScopes.Count; index++)
-        {
-            var (jsonScope, kind) = additionalScopes[index];
-
-            if (!seenScopes.Add(jsonScope))
-            {
-                continue;
-            }
-
-            normalized.Add((jsonScope, kind, CountScopeDepth(jsonScope), index));
-        }
-
-        normalized.Sort(
-            static (left, right) =>
-            {
-                var depthComparison = left.Depth.CompareTo(right.Depth);
-                return depthComparison != 0 ? depthComparison : left.Index.CompareTo(right.Index);
-            }
-        );
-
-        return [.. normalized.Select(scope => (scope.JsonScope, scope.Kind))];
-    }
-
-    private static int CountScopeDepth(string jsonScope) => jsonScope.Count(c => c == '.') + 1;
 }
