+++ src/dms/backend/EdFi.DataManagementService.Backend/Profile/ScopeTopologyIndex.cs
@@ -5,6 +5,7 @@
 
 using EdFi.DataManagementService.Backend.External;
 using EdFi.DataManagementService.Backend.External.Plans;
+using EdFi.DataManagementService.Backend.External.Profile;
 using EdFi.DataManagementService.Core.Profile;
 
 namespace EdFi.DataManagementService.Backend.Profile;
@@ -62,7 +63,7 @@ internal sealed class ScopeTopologyIndex
             .TablePlansInDependencyOrder.Select(tp => tp.TableModel.JsonScope.Canonical)
             .ToHashSet(StringComparer.Ordinal);
 
-        var normalizedAdditionalScopes = NormalizeAdditionalScopes(additionalScopes, tableScopes);
+        var normalizedAdditionalScopes = InlinedScopeNormalization.Normalize(additionalScopes, tableScopes);
 
         // First pass: collect all collection-kinded scopes (table-backed plus any inlined
         // collections the caller passes in) so nested detection sees them all.
@@ -155,48 +156,6 @@ internal sealed class ScopeTopologyIndex
 
         return false;
     }
-
-    /// <summary>
-    /// Normalizes caller-supplied inlined scopes so ancestors are processed before descendants.
-    /// Skips duplicate scopes and scopes already represented by a table-backed plan entry.
-    /// </summary>
-    private static IReadOnlyList<(string JsonScope, ScopeKind Kind)> NormalizeAdditionalScopes(
-        IReadOnlyList<(string JsonScope, ScopeKind Kind)>? additionalScopes,
-        IReadOnlySet<string> knownScopes
-    )
-    {
-        if (additionalScopes is not { Count: > 0 })
-        {
-            return [];
-        }
-
-        List<(string JsonScope, ScopeKind Kind, int Depth, int Index)> normalized = [];
-        HashSet<string> seenScopes = new(knownScopes, StringComparer.Ordinal);
-
-        for (var index = 0; index < additionalScopes.Count; index++)
-        {
-            var (jsonScope, kind) = additionalScopes[index];
-
-            if (!seenScopes.Add(jsonScope))
-            {
-                continue;
-            }
-
-            normalized.Add((jsonScope, kind, CountScopeDepth(jsonScope), index));
-        }
-
-        normalized.Sort(
-            static (left, right) =>
-            {
-                var depthComparison = left.Depth.CompareTo(right.Depth);
-                return depthComparison != 0 ? depthComparison : left.Index.CompareTo(right.Index);
-            }
-        );
-
-        return [.. normalized.Select(scope => (scope.JsonScope, scope.Kind))];
-    }
-
-    private static int CountScopeDepth(string jsonScope) => jsonScope.Count(c => c == '.') + 1;
 
     /// <summary>
     /// Walks the dot-separated ancestor prefixes of <paramref name="jsonScope"/> (excluding the
